move gpu driver config to build script, rearrange build options and beautify the script more.

+++ build_slmkernel.sh
@@ -28,35 +28,61 @@ rm -rf $CFGDIR/compiled_defconfig
 make -C $(pwd) O=$(pwd)/out clean -j$(nproc) && make -C $(pwd) O=$(pwd)/out mrproper -j$(nproc)
 clear
  
-read -p "`echo -e 'thanks for building slmkernel \ntell what device you wanna build for 💩💩 \nsupported devices: a22, a32, m32(experimental), f22(experimental), m22(very experimental)  '`" choice
+read -p "`echo -e 'thanks for building slmkernel \ntell what device you wanna build for 💩💩 \nsupported devices: a32, a22, f22, m22(experimental), m32(experimental)  '`" choice
 case "$choice" in 
-  a22|A22 ) export DEVICE="a22";;
   a32|A32 ) export DEVICE="a32";;
-  m32|M32 ) export DEVICE="m32";;
+  a22|A22 ) export DEVICE="a22";;
   f22|F22 ) export DEVICE="f22";;
   m22|M22 ) export DEVICE="m22";;
+  m32|M32 ) export DEVICE="m32";;
   * ) echo "u made a typo or $choice not supported yet srry 💩" && exit;;
 esac
 
 #edit perf.config to battery.config to disable perf tweaks, dont use them at the same time!
 #add $CFGDIR/ksu.config at the end before ">" for ksu integration(optional)
 #example: build m22 battery life oriented karnal with ksu: $CFGDIR/mt6768_slm_defconfig $CFGDIR/"$DEVICE".config $CFGDIR/battery.config $CFGDIR/ksu.config
-cat $CFGDIR/mt6768_slm_defconfig $CFGDIR/"$DEVICE".config $CFGDIR/perf.config > $CFGDIR/compiled_defconfig
+cat $CFGDIR/mt6768_slm_defconfig $CFGDIR/"$DEVICE".config $CFGDIR/battery.config > $CFGDIR/compiled_defconfig
 
-#selinux control it by here
-echo "
+#selinux and gpu driver control
+#buildable: mali bifrost r25p0, mali valhall r32p1, mali avalon r49p1[WIP]
+echo '
 # CONFIG_ALWAYS_ENFORCE is not set
 CONFIG_ALWAYS_PERMISSIVE=y
-" >> $CFGDIR/compiled_defconfig
+
+CONFIG_MTK_GPU_VERSION="mali valhall r32p1"
+' >> "$CFGDIR/compiled_defconfig"
 
 make -C $(pwd) O=$(pwd)/out -j$(nproc) compiled_defconfig
 make -s -C $(pwd) O=$(pwd)/out -j$(nproc)
-echo "u (tried to) built for: $DEVICE"
 
-#only for me delete if u want 💩💩💩💩
-read -p "copy to kernal directory? (are u vigus?) y/n   " choice
-case "$choice" in 
-  y|Y ) cp out/arch/arm64/boot/Image ~/Downloads/buildkernal/Image;;
-  n|N ) echo "k";;
-  * ) echo "nvm";;
-esac
+IMAGECHECK="$(pwd)/out/arch/arm64/boot/Image"
+
+if [ -f "$IMAGECHECK" ]; then
+    echo "built slm for device: $DEVICE"
+    GPU_VER=$(sed -n 's/^CONFIG_MTK_GPU_VERSION="\([^"]*\)"/\1/p' \
+        "$(pwd)/out/.config")
+
+    if [ "$GPU_VER" != "mali bifrost r25p0" ]; then
+        echo
+        echo "================================================================"
+        echo "warning: non-stock gpu driver selected"
+        echo
+        echo "built gpu driver : $GPU_VER"
+        echo
+        echo "Flash a custom vendor.img that has the corresponding Mali userspace libs,"
+        echo "Flash a custom boot.img that has a modified DTS with the new driver support,"
+        echo "or your device may bootloop"
+        echo "================================================================"
+        echo
+    fi
+
+    #only for me delete if u want 💩💩💩💩
+    read -p "copy to kernal directory? (are u vigus?) y/n   " choice
+    case "$choice" in 
+      y|Y ) cp out/arch/arm64/boot/Image ~/Downloads/buildkernal/Image;;
+      n|N ) echo "k";;
+      * ) echo "nvm";;
+    esac
+fi
+
+echo "$DEVICE"
